Factor storage clients into separate modules. Closes #1355.
* Refactor storage_clients.py from a script to a module inside chainlit.data
* Add pytest on S3StorageClient + add moto in tests dependencies
* Moved BaseStorageClient from from chainlit.data.base.py to chainlit.data.storage_clients.base.py
