+++ backend/chainlit/data/base.py
@@ -105,17 +105,3 @@ class BaseDataLayer(ABC):
     @abstractmethod
     async def build_debug_url(self) -> str:
         pass
-
-
-class BaseStorageClient(ABC):
-    """Base class for non-text data persistence like Azure Data Lake, S3, Google Storage, etc."""
-
-    @abstractmethod
-    async def upload_file(
-        self,
-        object_key: str,
-        data: Union[bytes, str],
-        mime: str = "application/octet-stream",
-        overwrite: bool = True,
-    ) -> Dict[str, Any]:
-        pass
