+++ backend/chainlit/data/storage_clients/base.py
@@ -1,0 +1,16 @@
+from abc import ABC, abstractmethod
+from typing import Any, Dict, Union
+
+
+class BaseStorageClient(ABC):
+    """Base class for non-text data persistence like Azure Data Lake, S3, Google Storage, etc."""
+
+    @abstractmethod
+    async def upload_file(
+        self,
+        object_key: str,
+        data: Union[bytes, str],
+        mime: str = "application/octet-stream",
+        overwrite: bool = True,
+    ) -> Dict[str, Any]:
+        pass
