+++ backend/pyproject.toml
@@ -80,6 +80,7 @@ slack_bolt = "^1.18.1"
 discord = "^2.3.2"
 botbuilder-core = "^4.15.0"
 aiosqlite = "^0.20.0"
+moto = "^5.0.14"
 
 [tool.poetry.group.dev.dependencies]
 black = "^24.8.0"
@@ -114,6 +115,7 @@ module = [
     "azure.storage.filedatalake",
 ]
 ignore_missing_imports = true
+
