+++ backend/tests/data/conftest.py
@@ -2,7 +2,7 @@ import pytest
 
 from unittest.mock import AsyncMock
 
-from chainlit.data.base import BaseStorageClient
+from chainlit.data.storage_clients.base import BaseStorageClient
 from chainlit.user import User
